+++ Userland/Libraries/LibJS/Runtime/Intl/DisplayNamesPrototype.cpp
@@ -6,7 +6,6 @@
 
 #include <AK/TypeCasts.h>
 #include <LibJS/Runtime/GlobalObject.h>
-#include <LibJS/Runtime/Intl/AbstractOperations.h>
 #include <LibJS/Runtime/Intl/DisplayNames.h>
 #include <LibJS/Runtime/Intl/DisplayNamesPrototype.h>
 #include <LibUnicode/Locale.h>
